Re-organized modules and refactored directory tree
