+++ src/WalletWave/plugins/wallet_scanner.py
@@ -1,8 +1,8 @@
+from WalletWave.config import ConfigManager
 from WalletWave.plugins.utils.plugin_interface import PluginInterface
 from WalletWave.repositories.gmgn_repo import GmgnRepo
 from WalletWave.utils.logging_utils import get_logger
-from WalletWave.config import ConfigManager
-import sys
+
 
 # Author: LetsStartWithPurple
 # Version: 1.0.1
