--- src/WalletWave/utils/gmgn_client/schemas/__init__.py
@@ -1,6 +1,0 @@
-from WalletWave.utils.gmgn_client.schemas.token_info import TokenInfoResponse
-from WalletWave.utils.gmgn_client.schemas.top_holders import TopHoldersResponse
-from WalletWave.utils.gmgn_client.schemas.wallets import WalletsResponse
-from WalletWave.utils.gmgn_client.schemas.wallet_info import WalletInfoResponse
-
-__all__ = ["TokenInfoResponse", "TopHoldersResponse", "WalletsResponse", "WalletInfoResponse"]
